[Refactor] 카테고리별 피드백 점수 JPQL 변경
