+++ src/main/java/com/example/SchoolLunchReport/domain/statistics/domain/feedback/service/FeedBackService.java
@@ -2,11 +2,11 @@ package com.example.SchoolLunchReport.domain.statistics.domain.feedback.service;
 
 import com.example.SchoolLunchReport.domain.product.menu.domain.entity.Menu;
 import com.example.SchoolLunchReport.domain.statistics.controller.dto.response.TrackingResponseDto;
-import com.example.SchoolLunchReport.domain.statistics.domain.feedback.entity.CategoryScoreAvgDto;
+import com.example.SchoolLunchReport.domain.statistics.domain.boundary.entity.Boundary;
 import com.example.SchoolLunchReport.domain.statistics.domain.feedback.entity.FeedBack;
+import com.example.SchoolLunchReport.domain.statistics.domain.feedback.repo.dto.CategoryScoreAvgDto;
 import com.example.SchoolLunchReport.domain.statistics.domain.feedback.support.FeedBackReader;
 import com.example.SchoolLunchReport.domain.statistics.domain.feedback.support.FeedBackTracker;
-import com.example.SchoolLunchReport.domain.statistics.domain.boundary.entity.Boundary;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
